uplded flwchart i gess
It is rare that such a humble user-editable box should appear to allow me to add a heavily extended description of my actions immediately past. It brings me great joy and satisfaction to be delivered such a means to passionately throw away my time in order to write said description. It is now that I can formally announce that I have indeed uploaded the flowchart, drawn to the best of my utility-utilising skills via the online land of draw.io, to the best of my upload-utilising skills. I would like to add, that I should hope I have done a satisfactory job; for while the flowchart took me virtually a negligible amount of my time, typing in this glorious extended description took much more, perhaps by a factor of infinity. It is even more likely that this is the only part of the assignment--within which this is in fact not a part, and as I reiterate is simply a passionate waste of my time--that will not be read by eyes other than my own. Nevertheless, let us bathe in the glory of the uploading of the flowchart by the title of Flowchart.png, and let us cheer; hip hip, hooray! Hip hip, hooray! In such a passionately British and over-complicated fashion; to celebrate only the most intriguing of uploads, the greatest of virtual imagery, the pure, programmer-centric symbolism of time so passionately passed by with no useful purposewhatsoever, let us cheer! But wait! The cheers! Where have they gone? The room is silent, for the celebratory phase has passed, as perhaps, just perhaps, has the lifetime of this extended description. But we cannot leave yet. We have not yet spoken our farewells... Farewell, glorious extended description box of the almighty Internet! Farewell, forgotten, lost time of the passionate programmer! And farewell you, the reader....which, erm, who are you, and why did you read oh so much? Hath thou beenst dragged into mine rabbit hole withest me? I salute you if thou hath, and if thou may please, say farewell with me! Farewell! Farewell! Until we make the grand mistake of typing into an extended description box again, farewell! And may you be committed as always!